enhance test runner to check for return codes
update the test runner to put the exit code in the output,
and update the expectation files to have the expected exit
code, so they are validated. There is a regression in one
of the tokenization tests that was missed since I didn't
have that. Now need to go fix the regression.